temporary remove eat() and move()
