+++ life/src/main/java/org/yt/jr/projects/creatures/animals/Deer.java
@@ -1,7 +1,7 @@
 package org.yt.jr.projects.creatures.animals;
 
 import org.yt.jr.projects.creatures.Animal;
-import org.yt.jr.projects.utils.CreaturesTypes;
+import org.yt.jr.projects.creatures.CreaturesTypes;
 
 public class Deer extends Animal {
 
@@ -9,13 +9,4 @@ public class Deer extends Animal {
         super(CreaturesTypes.DEER, health);
     }
 
-    @Override
-    public void eat() {
-
-    }
-
-    @Override
-    public void move() {
-
-    }
 }
