+++ life/src/main/java/org/yt/jr/projects/creatures/animals/Horse.java
@@ -1,7 +1,7 @@
 package org.yt.jr.projects.creatures.animals;
 
 import org.yt.jr.projects.creatures.Animal;
-import org.yt.jr.projects.utils.CreaturesTypes;
+import org.yt.jr.projects.creatures.CreaturesTypes;
 
 public class Horse extends Animal {
 
@@ -9,13 +9,4 @@ public class Horse extends Animal {
         super(CreaturesTypes.HORSE, health);
     }
 
-    @Override
-    public void eat() {
-
-    }
-
-    @Override
-    public void move() {
-
-    }
 }
